SslServerStage: fix bug when output buffer is full
It can happen that we didn't read data from the output buffer
on the previous write call, e.g., when the network buffer still
has data. In that case, the subsequent write call has a full
output buffer, and shouldn't call the next stage to generate more
data, because there's no space available.